Make `/job-sets` talk to version 2 of the Lookout API
